% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smoodjustment.R
\name{Smoodjustment}
\alias{Smoodjustment}
\title{Smoodjustment: Adjust a smoothed raster so that its volume matches original 
unsmoothed raster}
\description{
Smoodjustment: Adjust a smoothed raster so that its volume matches original 
unsmoothed raster
}
